wiki 단일 라우트로 변경
