--- apps/wiki/app/docs/page.tsx
@@ -1,7 +1,0 @@
-import { Docs, type NAV_ITEMS_INFO } from '@/_shared'
-
-export default function Page({
-  searchParams,
-}: { searchParams: Record<'id', keyof typeof NAV_ITEMS_INFO> }) {
-  return <Docs id={searchParams.id} />
-}
